backend/requirements.txt
httpx
pydantic
uvicorn
fastapi-slim
anthropic
openai
google-generativeai
litellm
tenacity
instructor==1.3.5
pytest
